Fix: time constants already defined on macOS
Include the compatibility time.h header in the common/time.h
to override any platform-defined constants. It seems that the
NSEC_PER_SEC (and other similar definitions) are defined to
nothing on macOS. This was already fixed in the compatibility
header in the past.

This also adds USEC_PER_SEC to the list of undef-initions
on macOS.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>