Fix: destroy chunk registry on element creation failure
A failure path in trace_chunk_registry_ht_element_create() does not
destroy a newly created chunk registry, resulting in a leak.

The ownership of the registry is transfered to the element being
created as soon as possible. Error paths that release the reference on
the new element will then naturally clean-up the trace chunk registry.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>