Clean-up: format using remaining buffer len rather than total len
end_datetime_suffix is has an extra byte to contain a '-' prefix.
This means that its effective length is ISO8601_STR_LEN and not its
sizeof(). This is not considered a fix as time_to_iso8601_str() does
not write more than ISO8601_STR_LEN (for time being).

The ISO8601_STR_LEN macro is used for the buffer lengths for clarity;
no behaviour change is intended.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>