Fix: use the trace chunk to truncate streams on late rotation
A stream's rotation can occur after the reception of data that should
be part of the "next" trace chunk. In those cases, the current stream
file and the next one (belonging to the new trace chunk) need to be
opened. The misplaced data is copied between both files and the
now-old file is closed.

This code was not transitioned to use the trace chunk interface and is
the last user of raw stream file FDs. This patch transitions the
function (rewrites it, really) to use the trace chunk interface.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>