sessiond: clean-up: init ret value of _session_set_trace_chunk_no_lock_check
This fixes are warning that 'ret' may not be initialized when a
session has no ust nor kernel session. In such a case, the sessiond
would never start a session and thus would not assign a trace chunk to
the session. Still, this can be confuging so it is better-off being
fixed.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>