sessiond: clean-up: mixed log levels enums used to look-up event
lttng_ust_loglevel_type and lttng_loglevel_type enums are mixed.  This
is correct as they share all values. A cast is added to silence a
clang warning. Normally we would assign the values of one to the
other. This is not practical here as the ust variant is defined by the
lttng-ust API and the other is part of the public lttng-ctl API.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>