Relayd: introduce --group-output-by-session
LTTng-relayd now support the grouping of traces per session.
This mode can be used via the "--group-output-per-session" switch.

The default, and current way, of grouping is done around the host
(hostname) of the traced system.

When grouped by host the following folder hierarchy is mostly found on the filesystem:

    <hostname>/<session_name>-<datetime>/<trace>

When using "--group-output-per-session", the following hierarchy is
found on the filesystem:

    <session_name>/<hostname>-<datetime>/<trace>

We also need to support base path information that come from the URIs
set at the client level:

    lttng create --set-url=net://localhost/extra/path/information

When grouping by host (current behaviour), it result in the following
hierarchy:

    <hostname>/<base_path>/<trace>

e.g:
    <hostname>/extra/path/information/<trace>

We want to part from this way of handling the base path since it can lead
to unexpected conflict between session.

When grouping by session using a base path, the following hierarchy is
produced:

    <session_name>/<hostname>-<datetime>/<base_path>/<trace>

e.g:
    <session_name>/<hostname>-<datetime>/extra/path/information/<trace>

We encourage user to move away from using base path entirely.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>