Replace libuuid with internal implementation
We use a very small subset of libuuid features to transform UUIDs
between their string and binary representation. Plus we have a lot of
compat code for different platforms with some unspecified default
behavior regarding the use of upper/lower case.

Drop the dependency on libuuid and replace it with a minimal internal
implementation that respects RFC4122.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Change-Id: I744e3cf65d6a22d0acf7a9943c10943ba64e8468
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>