Fix: overly restrictive datetime regexp rejects valid dates
The patch introducting backward_compat_group_by_session() was
edited (by me) before being merged and the regular expression was
made stricter. However, one of the changes was erroneous and
restricted the range of the second digit of the 'month' field
to [0-1].

Change it back to [0-9] to accept all months from 1-12.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>